refactor: mydashbaord page
